2 byte pin index #4097 (#4102)
* 2 byte pin index #4097

* 2 byte pin index #4097

good time to defragment configuration

* 2 byte pin index #4097

new fancy way

* 2 byte pin index #4097

smart comment

+++ firmware/controllers/algo/rusefi_hw_enums.h
@@ -20,7 +20,7 @@ typedef enum __attribute__ ((__packed__)) {
 /**
  * Hardware pin. This enum is platform-specific.
  */
-enum class Gpio : uint8_t {
+enum class Gpio : uint16_t {
 	Unassigned = 0,
 	// only used as return value of 'parseBrainPin' function do we really this this logic special value at all?!
 	Invalid = 1,
